% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermilleProgress.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Plot the progress of the alcohol concentration!}
\usage{
show_me_how_drunk(age, sex, height, weigth, drinking_time, drinks)
}
\arguments{
\item{age}{is a numeric integer for the age in years of a given person!}

\item{sex}{is a character variable, indicating the gender of the given person!}

\item{height}{is the size of the given person in cm!}

\item{weigth}{is a the weight of a given person in kg}

\item{drinking_time}{is a vector of dates, with the start and the end time!}

\item{drinks}{is a list of drinks, the person has drunk}
}
\value{
plot of the progress of the blood alcohol concentraion of a person!
}
\description{
used to  plot the permille of a person over the time, with 5 min intervalls!
}
