% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Get your approximate blood alcohol concentration!!!}
\usage{
tell_me_how_drunk(age, sex = c("male", "female"), height, weight,
  drinking_time, drinks)
}
\arguments{
\item{age}{is a numeric integer for the age in years of a given person!}

\item{sex}{is a character variable, indicating the gender of the given person!}

\item{height}{is the size of the given person in cm!}

\item{weight}{is a the weight of a given person in kg}

\item{drinking_time}{is a vector of dates, with the start and the end time!}

\item{drinks}{is a list of drinks, the person has drunk}
}
\value{
a numeric value of the approximate blood alcohol concentration of the
person at the end of the given drinking time
}
\description{
tell_me_how_drunk is used to calculate the alcohlic concentration in the 
blood of a person, given its attributes (weight, sex,...) according to 
[this page] ("https://thatsthefinger.com/")
}
